Drop device name comparison from "is-supported" check
The known device name string and the device name (as based on info
returned by the HID API and the device's firmware) are not in any
way related.  It just so happened that for my Ambit all these bits
of info lined up to match.  Based on that I mistakenly added this
to the "is-supported" check.  Sorry for any inconvenience caused.