Add `-u` option to ease post-build cleanup with Docker
Within a Docker container everything is run with `root` privileges so
that the build results will be owned by `root`.  As a result, you will
need `root` privileges to clean up.  The documentation now suggests to
use the `-u` option so that whoever ran the `docker` command can clean
up without elevated privileges.