Fix incorrect computing of number of packets to send
We can include 42 bytes in the first packet and then try to compute
how many more packets are needed, where each packet can take 54 bytes.

The "2" used for now (1 for the initial packet and 1 for the "remainder" data)
was not correct if (datalen - 42) can be divided by 54 with remainder 0.

So we should not add this final packet if datalen is a multiply of 54!

Solves #197